check errors in providers
