+++ internal/service/project/resource_project_migration_test.go
@@ -26,7 +26,7 @@ func TestAccMigrationProjectRS_NoProps(t *testing.T) {
 		CheckDestroy: acc.CheckDestroyProject,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config: fmt.Sprintf(`resource "mongodbatlas_project" "test" {
 					name   = "%s"
 					org_id = "%s"
@@ -82,7 +82,7 @@ func TestAccMigrationProjectRS_Teams(t *testing.T) {
 		CheckDestroy: acc.CheckDestroyProject,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            configWithTeams,
 				Check: resource.ComposeTestCheckFunc(
 					acc.CheckProjectExists(resourceName, &project),
@@ -121,7 +121,7 @@ func TestAccMigrationProjectRS_WithFalseDefaultSettings(t *testing.T) {
 		CheckDestroy: acc.CheckDestroyProject,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            configWithTeams,
 				Check: resource.ComposeTestCheckFunc(
 					acc.CheckProjectExists(resourceName, &project),
@@ -166,7 +166,7 @@ func TestAccMigrationProjectRS_WithLimits(t *testing.T) {
 		CheckDestroy: acc.CheckDestroyProject,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            config,
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "name", projectName),
@@ -203,7 +203,7 @@ func TestAccMigrationProjectRSProjectIPAccesslist_SettingIPAddress(t *testing.T)
 		CheckDestroy: acc.CheckDestroyProjectIPAccessList,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            acc.ConfigProjectIPAccessListWithIPAddress(orgID, projectName, ipAddress, comment),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(resourceName, "project_id"),
@@ -239,7 +239,7 @@ func TestAccMigrationProjectRSProjectIPAccessList_SettingCIDRBlock(t *testing.T)
 		CheckDestroy: acc.CheckDestroyProjectIPAccessList,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            acc.ConfigProjectIPAccessListWithCIDRBlock(orgID, projectName, cidrBlock, comment),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(resourceName, "project_id"),
@@ -294,7 +294,7 @@ func TestAccMigrationProjectRSProjectIPAccessList_Multiple_SettingMultiple(t *te
 		CheckDestroy: acc.CheckDestroyProjectIPAccessList,
 		Steps: []resource.TestStep{
 			{
-				ExternalProviders: mig.ExternalProviders(),
+				ExternalProviders: mig.ExternalProviders(t),
 				Config:            acc.ConfigProjectIPAccessListWithMultiple(projectName, orgID, accessList, false),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttrSet(resourceName, "project_id"),
